requirements.txt
python-lsp-server
python-can
cantools
canmatrix
opencv
